# javaclaw-api/javaclaw-api-admin/src/main/java/ai/javaclaw/api/admin/skills/SkillController.java
package ai.javaclaw.api.admin.skills;

import jakarta.validation.Valid;
import java.util.List;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/** CRUD endpoints for agent skills. */
@RestController
@RequestMapping("/api/skills")
public class SkillController {

    private final SkillService service;

    public SkillController(final SkillService service) {
        this.service = service;
    }

    @GetMapping
    public List<SkillDto> list() {
        return service.list();
    }

    @PostMapping
    public ResponseEntity<SkillDto> create(@Valid @RequestBody final SkillDto body) {
        return ResponseEntity.status(201).body(service.create(body));
    }

    @PutMapping("/{id}")
    public SkillDto update(@PathVariable final String id, @Valid @RequestBody final SkillDto body) {
        return service.update(id, body);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> delete(@PathVariable final String id) {
        service.delete(id);
        return ResponseEntity.noContent().build();
    }
}