# specs/openapi.yaml
openapi: 3.1.0
info:
  title: JavaClaw API
  version: 0.1.0
  description: >
    REST API for the JavaClaw AI platform.
    Covers chat (Vercel AI SDK v4 SSE streaming), conversation management,
    workspace file CRUD, MCP server registrations, agent skills, and system endpoints.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: Same-origin (default Spring Boot embedded server)
...
security:
  - basicAuth: []
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >
        HTTP Basic authentication. Currently unenforced (guest fallback active).
        Will be required starting Phase 4.1 (Spring Security hardening).

  # ── Reusable error responses ───────────────────────────────────────────────
  responses:
    Error400:
      description: Bad request — validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: Bad Request
            message: "content: must not be blank"
            timestamp: "2026-04-05T10:00:00Z"
    Error404:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: Not Found
            message: "Conversation abc123 not found"
            timestamp: "2026-04-05T10:00:00Z"
    Error429:
      description: Too many requests — SSE concurrency limit exhausted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: Too Many Requests
            message: "SSE concurrency limit reached"
            timestamp: "2026-04-05T10:00:00Z"
    Error500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: Internal Server Error
            message: "Unexpected error"
            timestamp: "2026-04-05T10:00:00Z"

  # ── DTO schemas (13 domain + 2 system = 15 total) ─────────────────────────
  schemas:
...
    ChatSendRequest:
      type: object
      description: >
        Request body for POST /api/chat/send.
        `content` is the user message text (required, 1–32000 chars).
        `conversationId` is optional; the server generates one when absent or blank.
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 32000
          description: User message text
        conversationId:
          type: string
          description: >
            Existing conversation ID to append to.
            Omit or send null/blank to start a new conversation.

    # ── Conversations ─────────────────────────────────────────────────────────
...
    FileNodeDto:
      type: object
      description: >
        Tree node returned by GET /api/files.
        `children` is present and non-null when `type` is "dir"; absent for files.
      required:
        - path
        - name
        - type
        - size
      properties:
        path:
          type: string
          description: Workspace-relative path (e.g. "src/Main.java")
        name:
          type: string
          description: File or directory name (last path segment)
        type:
          type: string
          enum:
            - file
            - dir
          description: Node type
        size:
          type: integer
          format: int64
          minimum: 0
          description: File size in bytes; 0 for directories
        children:
          type: array
          items:
            $ref: '#/components/schemas/FileNodeDto'
          description: Recursive child nodes; present only when type=dir
    FileContentDto:
      type: object
      description: File path and raw text content, used for both read and write responses.
      required:
        - path
        - content
      properties:
        path:
          type: string
          description: Workspace-relative file path
        content:
          type: string
          description: Raw UTF-8 text content of the file
    CreateFileRequest:
      type: object
      description: Body for POST /api/files — create a new file.
      required:
        - path
      properties:
        path:
          type: string
          minLength: 1
          description: Workspace-relative path for the new file (must not start with /)
        content:
          type: string
          description: Initial file content; empty string when omitted

    # ── MCP Servers ───────────────────────────────────────────────────────────
    McpServerDto:
      type: object
      description: MCP server registration (stdio or http transport).
      required:
        - name
        - transport
      properties:
        id:
          type: string
          description: Server-assigned identifier; omit on create
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Human-readable name for this MCP server
        transport:
          type: string
          enum:
            - stdio
            - http
          description: Transport protocol
        command:
          type: string
          description: Shell command to launch the server (stdio transport only)
        url:
          type: string
          description: HTTP endpoint URL (http transport only)
        headers:
          type: object
          additionalProperties:
            type: string
          description: Additional HTTP headers sent with every request (http transport only)
        enabled:
          type: boolean
          default: true
          description: Whether the server is active
    McpServerStatusDto:
      type: object
      description: Connection-state snapshot for an MCP server.
      required:
        - id
        - status
      properties:
        id:
          type: string
          description: MCP server identifier
        status:
          type: string
          description: >
            Connection status string, e.g. "connected", "disconnected", "error"
        detail:
          type: string
          description: Optional human-readable detail (error message, latency, etc.)

    # ── Skills ────────────────────────────────────────────────────────────────
    SkillDto:
      type: object
      description: Agent skill definition. `id` is generated server-side on create.
      required:
        - name
      properties:
        id:
          type: string
          description: Server-assigned identifier; omit on create
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Skill name
        description:
          type: string
          maxLength: 500
          description: Human-readable description of what the skill does
        enabled:
          type: boolean
          default: true
          description: Whether the skill is active

    # ── System ────────────────────────────────────────────────────────────────
...
    ApiError:
      type: object
      description: >
        Uniform error envelope returned by all /api/** error responses.
        Matches ai.javaclaw.api.admin.ApiError record.
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Short error classification (e.g. "Bad Request", "Not Found")
        message:
          type: string
          description: Human-readable error detail
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp of when the error occurred

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /api/chat/send:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message and receive a streaming response
      description: >
        Sends a user message to the AI agent and returns a **Vercel AI SDK v4
        data-stream** response body over a plain-text chunked connection.


        **Response content-type:** `text/plain;charset=UTF-8`


        **Response header:** `x-vercel-ai-data-stream: v1`


        **Frame format:** Each line is `<code>:<json>\n` where `<code>` is:

        - `0` — Text delta: incremental assistant text chunk (also used as keepalive heartbeat with `""`)

        - `2` — Data part: arbitrary JSON data part

        - `3` — Error: stream-level error text

        - `9` — Tool call: `{"toolCallId":"…","toolName":"…","args":{…}}`

        - `a` — Tool result: `{"toolCallId":"…","result":{…}}`

        - `d` — Finish message: `{"finishReason":"stop","usage":{"promptTokens":0,"completionTokens":0}}`

        - `e` — Finish step: `{"finishReason":"stop","usage":{…},"isContinued":false}`


        Returns **429** when the SSE concurrency semaphore is exhausted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSendRequest'
            example:
              content: "What is the capital of France?"
              conversationId: "web-abc123"
      responses:
        "200":
          description: SSE data stream — newline-delimited Vercel AI SDK v4 frames
          headers:
            x-vercel-ai-data-stream:
              description: Protocol version indicator consumed by @ai-sdk/react useChat
              schema:
                type: string
                enum: [v1]
          content:
            text/plain:
              schema:
                type: string
                description: Newline-delimited data-stream frames
              example: "0:\"Hello\"\n0:\" world\"\ne:{\"finishReason\":\"stop\",\"usage\":{\"promptTokens\":0,\"completionTokens\":0},\"isContinued\":false}\nd:{\"finishReason\":\"stop\",\"usage\":{\"promptTokens\":0,\"completionTokens\":0}}\n"
        "400":
          $ref: '#/components/responses/Error400'
        "429":
          $ref: '#/components/responses/Error429'
        "500":
          $ref: '#/components/responses/Error500'
      security: []
  /api/chat/stream/{conversationId}:
    post:
      operationId: reopenChatStream
      tags: [chat]
      summary: Reopen a bare data-stream channel for an existing conversation
      description: >
        Opens a new SSE data-stream connection tied to `conversationId` without
        sending a message. Intended for reconnecting after a network interruption.


        Same response format and headers as `sendChatMessage`.
        Returns **429** when the SSE concurrency limit is exhausted.
      parameters:
        - name: conversationId
          in: path
          required: true
          description: ID of the conversation to reopen
          schema:
            type: string
      responses:
        "200":
          description: Bare data-stream channel opened (no assistant turn initiated)
          headers:
            x-vercel-ai-data-stream:
              description: Protocol version indicator
              schema:
                type: string
                enum: [v1]
          content:
            text/plain:
              schema:
                type: string
        "429":
          $ref: '#/components/responses/Error429'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

  # ---------------------------------------------------------------------------
  # Conversations
  # ---------------------------------------------------------------------------
...
  /api/files:
    get:
      operationId: getFileTree
      tags: [files]
      summary: Get the workspace file tree
      description: Returns the full recursive directory tree rooted at the workspace root.
      responses:
        "200":
          description: Root node of the workspace file tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileNodeDto'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

    post:
      operationId: createFile
      tags: [files]
      summary: Create a new file in the workspace
      description: Creates a new file at the given path with optional initial content.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileRequest'
            example:
              path: "src/Hello.java"
              content: "public class Hello {}"
      responses:
        "201":
          description: File created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileContentDto'
        "400":
          $ref: '#/components/responses/Error400'
        "500":
          $ref: '#/components/responses/Error500'
      security: []
  /api/files/{path}:
    get:
      operationId: readFile
      tags: [files]
      summary: Read a workspace file
      description: >
        Returns the path and UTF-8 text content of the file at `{path}`.


        **Note:** The path parameter is matched by Spring MVC's `/**` AntPathMatcher,
        so it may contain forward slashes (e.g. `src/main/App.java`).
      parameters:
        - name: path
          in: path
          required: true
          description: >
            Workspace-relative file path — may contain forward slashes.
            Example: `src/main/java/App.java`
          schema:
            type: string
            pattern: '^(?!/).+'
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileContentDto'
        "404":
          $ref: '#/components/responses/Error404'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

    put:
      operationId: writeFile
      tags: [files]
      summary: Write (overwrite) a workspace file
      description: >
        Writes `content` to the file at `{path}`, creating it if it does not exist.
        Returns the updated file.
      parameters:
        - name: path
          in: path
          required: true
          description: Workspace-relative file path (may contain forward slashes)
          schema:
            type: string
            pattern: '^(?!/).+'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileContentDto'
      responses:
        "200":
          description: File written successfully; returns updated content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileContentDto'
        "400":
          $ref: '#/components/responses/Error400'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a workspace file or directory
      parameters:
        - name: path
          in: path
          required: true
          description: Workspace-relative path to delete (may contain forward slashes)
          schema:
            type: string
            pattern: '^(?!/).+'
      responses:
        "204":
          description: File deleted successfully
        "404":
          $ref: '#/components/responses/Error404'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

  # ---------------------------------------------------------------------------
  # MCP Servers
  # ---------------------------------------------------------------------------
  /api/mcp-servers:
    get:
      operationId: listMcpServers
      tags: [mcp-servers]
      summary: List all registered MCP servers
      responses:
        "200":
          description: Array of MCP server registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/McpServerDto'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

    post:
      operationId: createMcpServer
      tags: [mcp-servers]
      summary: Register a new MCP server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpServerDto'
            example:
              name: "My Python MCP"
              transport: "stdio"
              command: "python3 -m my_mcp_server"
              enabled: true
      responses:
        "201":
          description: MCP server created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpServerDto'
        "400":
          $ref: '#/components/responses/Error400'
        "500":
          $ref: '#/components/responses/Error500'
      security: []
  /api/mcp-servers/{id}:
    put:
      operationId: updateMcpServer
      tags: [mcp-servers]
      summary: Update an existing MCP server registration
      parameters:
        - name: id
          in: path
          required: true
          description: MCP server identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpServerDto'
      responses:
        "200":
          description: Updated MCP server registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpServerDto'
        "400":
          $ref: '#/components/responses/Error400'
        "404":
          $ref: '#/components/responses/Error404'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

    delete:
      operationId: deleteMcpServer
      tags: [mcp-servers]
      summary: Delete an MCP server registration
      parameters:
        - name: id
          in: path
          required: true
          description: MCP server identifier
          schema:
            type: string
      responses:
        "204":
          description: MCP server deleted
        "404":
          $ref: '#/components/responses/Error404'
        "500":
          $ref: '#/components/responses/Error500'
      security: []
  /api/mcp-servers/{id}/status:
    get:
      operationId: getMcpServerStatus
      tags: [mcp-servers]
      summary: Get the connection status of an MCP server
      parameters:
        - name: id
          in: path
          required: true
          description: MCP server identifier
          schema:
            type: string
      responses:
        "200":
          description: Connection-state snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpServerStatusDto'
        "404":
          $ref: '#/components/responses/Error404'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

  # ---------------------------------------------------------------------------
  # Skills
  # ---------------------------------------------------------------------------
  /api/skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all agent skills
      responses:
        "200":
          description: Array of skill definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillDto'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a new agent skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillDto'
            example:
              name: "Web Search"
              description: "Search the web using DuckDuckGo"
              enabled: true
      responses:
        "201":
          description: Skill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDto'
        "400":
          $ref: '#/components/responses/Error400'
        "500":
          $ref: '#/components/responses/Error500'
      security: []
  /api/skills/{id}:
    put:
      operationId: updateSkill
      tags: [skills]
      summary: Update an existing skill
      parameters:
        - name: id
          in: path
          required: true
          description: Skill identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillDto'
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDto'
        "400":
          $ref: '#/components/responses/Error400'
        "404":
          $ref: '#/components/responses/Error404'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - name: id
          in: path
          required: true
          description: Skill identifier
          schema:
            type: string
      responses:
        "204":
          description: Skill deleted
        "404":
          $ref: '#/components/responses/Error404'
        "500":
          $ref: '#/components/responses/Error500'
      security: []

  # ---------------------------------------------------------------------------
  # System
  # ---------------------------------------------------------------------------
...